# feat(ScheduleController): smaller schedules_for_stop response + logging