--- a/apps/site/lib/site_web/controllers/schedule_controller.ex
+++ b/apps/site/lib/site_web/controllers/schedule_controller.ex
@@ -3,6 +3,8 @@ defmodule SiteWeb.ScheduleController do
   alias Routes.Route
   alias Schedules.Schedule
 
+  require Logger
+
   plug(SiteWeb.Plugs.Route when action not in [:schedules_for_stop])
 
   @spec show(Plug.Conn.t(), map) :: Phoenix.HTML.Safe.t()
@@ -47,20 +49,30 @@ defmodule SiteWeb.ScheduleController do
 
   @spec schedules_for_stop(Plug.Conn.t(), map) :: Plug.Conn.t()
   def schedules_for_stop(conn, %{"stop_id" => stop_id} = params) do
-    schedules =
-      Schedules.Repo.schedules_for_stop(stop_id, [])
-      |> future_departures(conn, params)
-      |> omit_last_stop_departures(params)
+    case Schedules.Repo.schedules_for_stop(stop_id, []) do
+      {:error, error} ->
+        _ =
+          Logger.error(
+            "module=#{__MODULE__} fun=schedules_for_stop stop=#{stop_id} date_time=#{DateTime.to_string(date_time(conn.assigns))} error=#{inspect(error)}"
+          )
 
-    json(conn, schedules)
+        SiteWeb.ControllerHelpers.return_internal_error(conn)
+
+      data ->
+        schedules =
+          data
+          |> future_departures(conn, params)
+          |> omit_last_stop_departures(params)
+
+        json(conn, schedules)
+    end
   end
 
+  defp date_time(%{"date_time" => date_time}), do: date_time
+  defp date_time(_), do: Util.now()
+
   defp future_departures(schedules, conn, %{"future_departures" => "true"}) do
-    now =
-      case conn.assigns do
-        %{"date_time" => date_time} -> date_time
-        _ -> Util.now()
-      end
+    now = date_time(conn.assigns)
 
     # Only list schedules with departure_time in the future
     schedules