#### Summary of changes

<!-- Link to relevant Asana task; remove if not applicable -->
**Asana Ticket:** [Stop pages intermittently missing Commuter Rail schedules](https://app.asana.com/0/385363666817452/1205814823003719/f)

I happened to be able to consistently reproduce the problem for a particular bus route on a particular stop on prod... but ultimately I never figured out a definitive cause for this. Splunk didn't reveal any errors. No network problems or console errors. I suspect that caching _might_ exacerbate the problem and cause the backend to return `[]` instead of re-trying within the cache `max_age` window. But still, ideally it wouldn't erroneously return `[]` in the first place.

This PR does 3 things:

1. Returns an error response if the `Schedules.Repo.schedules_for_stop/2` call returns `{:error, error}` value. An error will be logged in Splunk and the frontend would handle it by showing the error view with the bus gif.
2. Logs a message if the list of schedules returned happens to be empty. Just for informational/future debugging purposes should be issue persist and not be directly related to a `Schedules.Repo` error.
3. Dramatically reduces the size of the JSON response for `SiteWeb.ScheduleController.schedules_for_stop/2` in hopes of improving performance. It removes data that's not used on the frontend. 

Comparison for **Alewife Station** (`/api/stops/place-alfcl/schedules?last_stop_departures=false&future_departures=true`), using Chrome DevTools with no caching and "Fast 3G" network throttling setting:

|  | Before | After | **Difference**
| --- | --- | --- | --- |
|  Response size | 272 kB | 53.4 kB | **-80%** |
| Time | 2.49s | 885ms | **-64%** |


---

#### General checks
* [x] **Are the changes organized into self-contained commits with descriptive and well-formatted commit messages?** This is a good practice that can facilitate easier reviews.
* [x] **Testing.** Do the changes include relevant passing updates to tests? This includes updating screenshots. Preferably tests are run locally to verify that there are no test failures created by these changes, before opening a PR.
* [x] **Tech debt.** Have you checked for tech debt you can address in the area you're working in? This can be a good time to address small issues, or create Asana tickets for larger issues.